add details
