+++ components/StyledFlag.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const StyledFlag = styled.aside`
+  font-size: 3rem;
+`;
+
+export default StyledFlag;
